dans dépenses : int devenu double / dans leastSqueares : modification pour convenir à nos besoins / dans courbe : tracer tout !!

+++ Depenses.java
@@ -4,7 +4,7 @@ public class Depenses {
 	double montant;
 	String type;
 	String description;
-	int date;
+	double date;
 	
 	public Depenses (double amount, String categorie, String note){
 		montant = amount;
@@ -13,13 +13,13 @@ public class Depenses {
 		date= this.convertDate();
 	}
 	
-	public int convertDate(){
-		int num =0;
+	public double convertDate(){
+		double num =0;
 		Calendar cal = new GregorianCalendar(); 
-		int year = cal.get(Calendar.YEAR);             // 2005
-		int month = cal.get(Calendar.MONTH);           // 0=Jan, 1=Feb, ...
-		int day = cal.get(Calendar.DAY_OF_MONTH);      // 1...
-		int time = cal.get(Calendar.HOUR_OF_DAY);
+		double year = cal.get(Calendar.YEAR);             // 2005
+		double month = cal.get(Calendar.MONTH);           // 0=Jan, 1=Feb, ...
+		double day = cal.get(Calendar.DAY_OF_MONTH);      // 1...
+		double time = cal.get(Calendar.HOUR_OF_DAY);
 		num +=(2019-year) + month + day +time;
 		return num;
 	}
@@ -28,7 +28,7 @@ public class Depenses {
 		return this.montant;
 	}
 	
-	public int getDate(){
+	public double getDate(){
 		return this.date;
 	}
